+++ packages/bluesky/lib/src/services/codegen/app/bsky/graph/muteActor/input.g.dart
@@ -12,6 +12,8 @@ _GraphMuteActorInput _$GraphMuteActorInputFromJson(Map json) =>
     $checkedCreate('_GraphMuteActorInput', json, ($checkedConvert) {
       final val = _GraphMuteActorInput(
         actor: $checkedConvert('actor', (v) => v as String),
+        onlyReposts: $checkedConvert('onlyReposts', (v) => v as bool?),
+        onlyQuoteposts: $checkedConvert('onlyQuoteposts', (v) => v as bool?),
         $unknown: $checkedConvert(
           r'$unknown',
           (v) => (v as Map?)?.map((k, e) => MapEntry(k as String, e)),
@@ -24,5 +26,7 @@ Map<String, dynamic> _$GraphMuteActorInputToJson(
   _GraphMuteActorInput instance,
 ) => <String, dynamic>{
   'actor': instance.actor,
+  'onlyReposts': ?instance.onlyReposts,
+  'onlyQuoteposts': ?instance.onlyQuoteposts,
   r'$unknown': ?instance.$unknown,
 };